--- models/Userinfo.php
@@ -1,50 +1,0 @@
-<?php
-
-namespace app\models;
-
-use Yii;
-
-/**
- * This is the model class for table "userinfo".
- *
- * @property integer $id
- * @property string $username
- * @property string $email
- * @property string $password
- */
-class Userinfo extends \yii\db\ActiveRecord
-{
-    
-    /**
-     * @inheritdoc
-     */
-    public static function tableName()
-    {
-        return 'userinfo';
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function rules()
-    {
-        return [
-            [['username', 'email', 'password'], 'required'],
-            [['username', 'password'], 'string', 'max' => 20],
-            [['email'], 'string', 'max' => 100]
-        ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function attributeLabels()
-    {
-        return [
-            'id' => 'ID',
-            'username' => 'Username',
-            'email' => 'Email',
-            'password' => 'Password',
-        ];
-    }
-}
