--- views/userinfo/update.php
@@ -1,21 +1,0 @@
-<?php
-
-use yii\helpers\Html;
-
-/* @var $this yii\web\View */
-/* @var $model app\models\Userinfo */
-
-$this->title = 'Update Userinfo: ' . ' ' . $model->id;
-$this->params['breadcrumbs'][] = ['label' => 'Userinfos', 'url' => ['index']];
-$this->params['breadcrumbs'][] = ['label' => $model->id, 'url' => ['view', 'id' => $model->id]];
-$this->params['breadcrumbs'][] = 'Update';
-?>
-<div class="userinfo-update">
-
-    <h1><?= Html::encode($this->title) ?></h1>
-
-    <?= $this->render('_form', [
-        'model' => $model,
-    ]) ?>
-
-</div>
